aulas/assinatura_digital/requirements.txt
# bibliotecas
streamlit==1.39.0
pycryptodome==3.21.0
